x2X-coders-main/requirements.txt
Flask
python-dotenv
langchain
langchain-nvidia-ai-endpoints
langchain-community
faiss-cpu
openai
streamlit
langchain
llama-index
chromadb
openai
fastapi
uvicorn
python-dotenv
langchain
langchain-core
langchain-community
langchain-nvidia-ai-endpoints

